refactor margin
