+++ src/frame/XYFrame.js
@@ -15,6 +15,7 @@ import generateFrameTitle from '../svg/frameFunctions/generateFrameTitle';
 import SpanOrDiv from '../utils/SpanOrDiv';
 import VisualizationLayer from '../layers/VisualizationLayer/VisualizationLayer';
 import { adjustedPositionSize } from '../data/dataFunctions';
+import toMarginGraphic from "../svg/frameFunctions/toMarginGraphic";
 
 const getCanvasScale = context => {
   const devicePixelRatio = window.devicePixelRatio || 1;
@@ -114,6 +115,18 @@ const XYFrame = props => {
           right: baseMargin
         }
       : Object.assign({ top: 0, bottom: 0, left: 0, right: 0 }, baseMargin);
+
+
+  //todo: remove
+  const marginGraphic = toMarginGraphic({matte, size, margin, name});
+  const axes = [];
+  const canvasPostProcess = null;
+  const projectedCoordinateNames = {};
+  const renderPipeline = {};
+  const renderOrder = [];
+  const xScale = null;
+  const yScale = null;
+  const data =  [];
 
   const { adjustedPosition, adjustedSize } = adjustedPositionSize({
     size: [width, height],
